pyproject.toml
[project]
name = "reddit-mcp-poc"
version = "0.1.0"
description = "A minimal Model Context Protocol server for Reddit content access"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "asyncpraw>=7.8.1",
    "fastmcp>=0.8.0",
    "openai-agents>=0.2.8",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
]

[project.scripts]
reddit-mcp = "src.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
